Remove get_complete_position; default is bug-free
Additionally, fixes a linting error associated with mutating the
input_pattern string. Mutating that string seems unnecessary to me, and
it requires reading the Deoplete source code to understand. This way,
we're guaranteed that the regex represents the only input pattern we
care about (it was adding to an empty string before, so this preserves
current behavior).