Wrap notify::Watcher to handle vim's write-via-rename
The language server can request a watch on a file that is currently open
in vim. If it does so, LanguageClient-neovim adds a file system watcher
on that file. Doing this directly via `notify::Watcher` as it did
before, however, goes wrong on systems where a file watch is really an
inode watch, and vim writes to files by writing to a temporary file and
then renaming over the original file. These systems include Linux and
macOS.